Minor cleanup and move to WcsNDMap
